pyproject.toml
[project]
name = "img-caption"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.8",
    "python-dotenv>=1.1.1",
    "replicate>=1.0.7",
]
